Introducing Output Caching for Localization Endpoints
